requirements.txt
django
djangorestframework
django-model-utils
pymysql
django-rest-swagger
